+++ src/config.lua
@@ -1,18 +1,15 @@
 local config = {
-  enabled = true;
-  primary = "WeaponStaffSwing";
-  secondary = "WeaponStaffSwing";
-  aspect = "BaseStaffAspect_Secondary";
-  last_primary = "WeaponStaffSwing";
-  last_secondary = "WeaponStaffSwing";
-  last_aspect = "None";
+  last_primary = "MovedToSaveData";
+  last_secondary = "MovedToSaveData";
+  last_aspect = "MovedToSaveData",
   random_fusion_each_run = false;
 }
 
 local configDesc = {
-  last_primary = "DO NOT TOUCH";
-  last_secondary = "DO NOT TOUCH";
-  last_aspect = "DO NOT TOUCH";
+  random_fusion_each_run = "Fuse two random weapons and their aspects before the start of each run";
+  last_primary = "Obsolete, settings tied to save data instead";
+  last_secondary = "Obsolete, settings tied to save data instead";
+  last_aspect = "Obsolete, settings tied to save data instead";
 }
 
 return config, configDesc